Remove unnecessary laser methods
Move single-line update and draw instructions to LaserPool methods.
Combine laser-related structs into a single file.
