--- src/laser/mod.rs
@@ -1,2 +1,0 @@
-mod laser;
-pub(crate) mod laser_pool;
